Add BasicGun and Projectile weapon system
Introduce a basic ranged weapon and projectile system: add Circle sprite and import, create a Projectile scene (Area2D with collision and sprite) and corresponding Projectile.cs to handle movement, lifetime, collisions and deal damage. Add BasicGun.cs which finds the nearest enemy, instantiates projectiles, and enforces a fire rate plus a per-frame shot cap to avoid backlog on low frame rates. Wire the BasicGun and Projectile scene into Player.tscn. Update enemy API to support projectile damage: change BaseEnemy.TakeDamage to accept float and be virtual, and change EnemyStats.MaxHealth and ContactDamage to floats to allow fractional damage values.

+++ Scripts/Resources/EnemyStats.cs
@@ -6,9 +6,9 @@ namespace NeonSwarm.Resources;
 public partial class EnemyStats : Resource
 {
     [Export] public string EnemyName { get; set; } = "Enemy"; // The name of the enemy. Used for game over screen, score display and debugging.
-    [Export] public int MaxHealth { get; set; } = 3; // The maximum health of the enemy. When health reaches 0, the enemy dies.
+    [Export] public float MaxHealth { get; set; } = 3; // The maximum health of the enemy. When health reaches 0, the enemy dies.
     [Export] public float MoveSpeed { get; set; } = 200.0f; // The movement speed of the enemy. Higher values make the enemy move faster.
-    [Export] public int ContactDamage { get; set; } = 1; // The amount of damage the enemy deals to the player on contact. Hihher values mean more damage to the player.
+    [Export] public float ContactDamage { get; set; } = 1; // The amount of damage the enemy deals to the player on contact. Hihher values mean more damage to the player.
     [Export] public int ScoreValue { get; set; } = 10; // The amount of points the player gets for defeating this enemy.
 
     [Export] public Color BodyColor { get; set; } = Colors.Red; // The color of the enemy's body and glow.
